product_microservice/requirements.txt
flask
google-cloud-storage
pandas==1.0.5
numpy==1.19.1
sqlalchemy==1.3.18
sqlalchemy-utils